Created user class, working registration
