+++ cloud/main.js
@@ -4,7 +4,7 @@ var global = require('cloud/globals.js');
 
 // -- Experts -- //
 var phone_handler = require('cloud/phone.js');
-var user_handler = require('cloud/user.js');
+var user_handler = require('cloud/user_handler.js');
 var expert_handler = require('cloud/expert.js');
 
 
@@ -31,12 +31,10 @@ Parse.Cloud.define("incomingSMS", function(request, response) {
 Parse.Cloud.define("newUser", function(request, response) {
   console.log("---");
 
-  var user = user_handler.makeUser("gefthefrench@gmail.com", "Jeff", "French", "3128602305", "https://dl.dropboxusercontent.com/u/30415492/Resume_Geoffroy.pdf");
+  var user = user_handler.makeAUser("gefthefrench@gmail.com", "Jeff", "French", "3128602305", "https://dl.dropboxusercontent.com/u/30415492/Resume_Geoffroy.pdf");
 
   if (!user)  response.error();
   else        response.success();
-
-  console.log("User: " + user);
 
   if (user) user.register();
